backend/requirements.txt
aiofiles==25.1.0
email-validator==2.1.1
fastapi==0.120.0
jinja2==3.1.4
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
Pillow==12.0.0
psycopg2-binary==2.9.9
pydantic==2.12.3
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
qrcode==8.2
redis==5.0.1
reportlab==4.4.4
Requests==2.32.5
schedule==1.2.2
sendgrid==6.12.5
SQLAlchemy==2.0.44
uvicorn==0.38.0
